Completely replace test_conn & test_core w/ pytest
